requirements.txt
customtkinter==5.2.1
hls4ml==0.8.0
keras==2.15.0
numpy==1.26.2
openlane==2.0.0b16
scikit_learn==1.3.2
tensorflow==2.15.0
tensorflow_model_optimization==0.7.5
tensor==0.3.6
q==2.7
qkeras==0.9.0
onnx==1.15.0
graphviz==0.20.1
pydot==1.4.2
seaborn==0.13.0
torch==2.1.1
tvm==1.0.0
volare==0.14.1
h5py==3.10.0
matplotlib==3.8.2
onnx==1.15.0
pandas==2.1.4
